CNI wrapper, new and improved!
Changed the name of the CNI_WRAPPER class to CniWrapper (the filename was also
updated to match the new class name).
Converted the class to function as a Singleton instead of being accessible
through a global pointer.
Added functions to simplify access to the conversion methods and configured the
header file to allow the CniWrapper to be used from both C and C++ programs.
Replaced leading spaces with TABs.