Updated to use/reflect the new and improved CNI wrapper.
Note that the scripts no longer explicitly create a CniWrapper object. It is
done indirectly during the first call to either cc2js() or js2cc(). The
ability to create it directly remains in the event that an instance is
necessary, but the conversion methods/functions are not actually used.