Added improved? implementation of toString() for public derived classes.
Rather than "dumping" a text representation of the entire instance, it
generates and returns an abbreviated description, which includes some
attributes. It is intended to model the basic information that appears on a
counter (e.g. for a Squad, it might be "6-2-8 Squad (A)" (the value in
parentheses is the ID, appearing only if one has been set)).