Changed Status.status() to return a List instead of a String.
Updates include modification of class implementing the interface (Fighting).
Note that the isStatusSet() method was removed, the thought being that it
would be just as easy to use the List to make the determination. For the
other (interface) methods, actual (i.e. more than previous placeholder)
implementations were added in Fighting.java