Added entries for the name() and toString() methods.
The static/immutable entries for each enum value have been removed since a)
they don't work in both scripting languages and b) it is not anticipated
that they would be used a lot (e.g. they are not passed to a constructor
like the other enums). The return type of valueOf was also changed to a
States* object to allow access to it via the other methods. It is also the
same type returned by the status() method (see Squad for example).