Added typemaps specific to handling States objects.
The first (in) translates a States* object into the type it wraps
(::jasl::counters::Status$States*) to allow it to be used in the library
code. The second typemap (out) translates the items in a ::java::util::List*
into an array of States objects. This currently works for Perl, but an
implementation for Python is forthcoming.