Updates to restore build and execution support for/with gcj.
The latest? version of GCC that supports gcj is 6.5.0. There is now a
dependency on a local build of that version (or one close to it) in (by
default) $JASL_BASE/tools. Changes were also made to support building the
JNI wrapper with a different (i.e., newer) version of GCC (it does work
with 6.5.0 as well).

The Python code was updated to support Python 3.