Updated to remove references to specific enums.
This was done as the first step in changing this interface so that it can
be used outside of the counters package / namespace. The documentation was
also improved. This includes removing external hyperlinks, which are
always out of date.

TODO: Move the toClass() function to the JniWrapper, as it is not specific
      to enums.